Got online gameplay working
Sending/receiving moves and interpreting them works just fine as it used to, but there are integration issues with threejs and react when strict mode is enabled.  Temporary solution is obviously to turn off strict mode, but this does mean there is some leakage of state.

Issue that made this commit so tedious is that when moves are received over the network, they sometimes do not get rendered to the correct scene (there should only be one).  Therefore, it would look like the opponent had not made their turn, but this was not a networking issue.

The long-term fix will either to completely re-write the component system into something like @react-three/fiber, or to be very careful about the application of stateful classes within components.