Implement tile placement in wasm
The placement of tiles in wasm with C is more or less complete, with
tests manually translated from the go tests. Uses GNU make and CMake to
build. This shouldn't be a problem with a one-click build if we just
include the wasm binary in the repository at some stage.