Get go server working and calling C code
Go comes with a c compiler, so the only way to truly call a shared
library is to first link with some C code that calls the shared library
and learn the cgo preprocessor commands, which is way more effort than
just including the main hive.c file into the go code and calling into
that.

Absolutely not ready to run with the new structure; just got it to work
without compiler/obvious runtime errors.