Generate normal maps for bugs and compress
Generate compressed normal maps to be sent over an internet connection.
Many of the old files for doing this were over 300KiB, but have now beed
reduced to ~25KiB each, with minimal difference in quality. This was
accomplished by resizing the images and applying compression where
visible differences are minimal. Inspecting the images closely, one
might notice that the normal maps are encodable to vector formats
(potentially with gradients where there are curves in geometry) so this
remains to be a future optimisation.

The build process for this has been documented in the `README.md` and is
managable with `norm.sh` for generating normal maps and `comp.sh` for
compressing these.