Get normal maps working
Note there is a slight hack to add a bit in the z (normal) direction and
then normalize in the shader because the particular normal maps we use
only point either fully in the tangent or normal direction.

Eventually, this should be fixed by updating the normal maps to point in
a less extreme direction.